refactor(server): extract query-validator to break compiler↔executor cycle (#851 PR1)
Move `validateQueryAgainstCubes` and its helpers (`getActiveValidationModes`,
`validateFilter`, `extractCubeFromRetentionTimeDimension`) out of compiler.ts
into a new standalone `src/server/query-validator.ts`. The executor now imports
the validator directly instead of from the compiler, breaking the
`compiler.ts → executor.ts → compiler.ts` circular dependency reported in #851.

compiler.ts re-exports `validateQueryAgainstCubes` for backward compatibility.

- madge: direct compiler↔executor cycle removed (15 → 14 cycles)
- typecheck clean; 2348 server tests pass (postgres)
