+++ src/server/executor.ts
@@ -31,7 +31,7 @@ import { generateCacheKey } from './cache-utils'
 import { DrizzleSqlBuilder } from './physical-plan/drizzle-sql-builder'
 import { LogicalPlanner } from './logical-plan/logical-planner'
 import { CTEBuilder } from './builders/cte-builder'
-import { validateQueryAgainstCubes } from './compiler'
+import { validateQueryAgainstCubes } from './query-validator'
 import { applyGapFilling } from './gap-filler'
 import type { DatabaseAdapter } from './adapters/base-adapter'
 import { ComparisonQueryBuilder } from './builders/comparison-query-builder'
